+++ 05_フロントエンド/02_適切にコンポーネントを分割して1ページ作ってみよう/app/src/components/molecule/todo-item/styles.module.css
@@ -1,0 +1,3 @@
+.completed {
+  background-color: gray;
+}
