+++ 05_フロントエンド/02_適切にコンポーネントを分割して1ページ作ってみよう/app/src/domain/models/data.ts
@@ -1,0 +1,5 @@
+export type Data = {
+  id: string;
+  task: string;
+  complete: boolean;
+};
