+++ 05_フロントエンド/02_適切にコンポーネントを分割して1ページ作ってみよう/app/src/domain/services/generate-id.ts
@@ -1,0 +1,2 @@
+export const generateID = (): string =>
+  (Math.floor(Math.random() * 90000) + 10000).toString();
